convo experiments

+++ trainers/dataloader.py
@@ -162,7 +162,7 @@ class ConversationalDataloader:
             data = np.memmap(
                 os.path.join(self.dataset_path, f"{split}.bin"), dtype=np.uint16, mode="r"
             )
-            self.load_shape = (len(data)/2/self.context_window, 2, self.context_window)
+            self.load_shape = (int(len(data)/2/self.context_window), 2, self.context_window)
             data = None 
         data = np.memmap(
             os.path.join(self.dataset_path, f"{split}.bin"),
